refactor(gexcel): rename excel to gexcel and finalize v2 merge
- Rename directory excel/ -> gexcel/
- Update package declarations from excel to gexcel
- Update import paths from github.com/morehao/golib/excel to
  github.com/morehao/golib/gexcel
- Merge v2 refactor (generic ReadFile/WriteFile, reader/writer core,
  schema parser, options, error model, new excel tag syntax)
- Update README docs and example to use new API
