+++ gexcel/api_read.go
@@ -1,4 +1,4 @@
-package excel
+package gexcel
 
 import "github.com/xuri/excelize/v2"
