+++ gexcel/api_read_test.go
@@ -1,4 +1,4 @@
-package excel
+package gexcel
 
 import (
 	"path/filepath"
